move stuff around to ut stuff is in ut module, and self review
